day 7 : update modularization
