+++ app/src/main/java/com/example/moviecollection/core/MainActivity.kt
@@ -7,7 +7,8 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
-import com.example.moviecollection.core.theme.MovieCollectionTheme
+import com.example.ui.MovieCollectionApp
+import com.example.ui.theme.MovieCollectionTheme
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
