--- app/src/main/java/com/example/moviecollection/core/component/CompLoading.kt
@@ -1,37 +1,0 @@
-package com.example.moviecollection.core.component
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.aspectRatio
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import com.airbnb.lottie.compose.LottieAnimation
-import com.airbnb.lottie.compose.LottieCompositionSpec
-import com.airbnb.lottie.compose.LottieConstants
-import com.airbnb.lottie.compose.rememberLottieComposition
-import com.example.moviecollection.R
-
-@Composable
-fun CompLoading(
-    modifier: Modifier = Modifier
-) {
-    val composition by rememberLottieComposition(spec = LottieCompositionSpec.RawRes(R.raw.anim_loading))
-
-    Column(
-        modifier.fillMaxSize(),
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Spacer(modifier = Modifier.weight(4f))
-        LottieAnimation(
-            modifier = Modifier
-                .aspectRatio(1f)
-                .weight(2f),
-            composition = composition,
-            iterations = LottieConstants.IterateForever
-        )
-        Spacer(modifier = Modifier.weight(4f))
-    }
-}
