--- app/src/main/java/com/example/moviecollection/ui/home/HomeState.kt
@@ -1,6 +1,0 @@
-package com.example.moviecollection.ui.home
-
-data class HomeState(
-    val isLoading: Boolean = true,
-    val errorMessage: String = ""
-)
