+++ settings.gradle.kts
@@ -18,3 +18,4 @@ include(":app")
 include(":data")
 include(":domain")
 include(":util")
+include(":ui")
