+++ ui/build.gradle.kts
@@ -2,11 +2,11 @@ plugins {
     id("com.android.library")
     id("org.jetbrains.kotlin.android")
     id("kotlin-kapt")
-    id("com.google.devtools.ksp")
+    id("com.google.dagger.hilt.android")
 }
 
 android {
-    namespace = "com.example.domain"
+    namespace = "com.example.ui"
     compileSdk = 34
 
     defaultConfig {
@@ -34,16 +34,44 @@ android {
         freeCompilerArgs = freeCompilerArgs +
                 "-opt-in=androidx.paging.ExperimentalPagingApi" +
                 "-opt-in=androidx.compose.material3.ExperimentalMaterial3Api" +
-                "-opt-in=androidx.compose.foundation.ExperimentalFoundationApi"
+                "-opt-in=androidx.compose.foundation.ExperimentalFoundationApi" +
+                "-opt-in=androidx.compose.material.ExperimentalMaterialApi"
+    }
+    buildFeatures {
+        compose = true
+    }
+    composeOptions {
+        kotlinCompilerExtensionVersion = "1.5.3"
+    }
+    packagingOptions {
+        resources {
+            excludes += "/META-INF/{AL2.0,LGPL2.1}"
+        }
     }
 }
 
 dependencies {
 
-    implementation("androidx.core:core-ktx:1.12.0")
-    implementation(platform("org.jetbrains.kotlin:kotlin-bom:1.8.0"))
     implementation("androidx.appcompat:appcompat:1.6.1")
-    implementation("com.google.android.material:material:1.11.0")
+
+    implementation("androidx.core:core-ktx:1.8.0")
+    implementation("androidx.lifecycle:lifecycle-runtime-ktx:2.7.0")
+    implementation("androidx.activity:activity-compose:1.8.2")
+    implementation(platform("androidx.compose:compose-bom:2023.08.00"))
+    implementation("androidx.compose.ui:ui")
+    implementation("androidx.compose.ui:ui-graphics")
+    implementation("androidx.compose.ui:ui-tooling-preview")
+    implementation("androidx.compose.material3:material3")
+    implementation("com.google.code.gson:gson:2.10")
+
+    implementation("androidx.compose.material:material:1.6.1")
+
+    //Coil
+    implementation("io.coil-kt:coil-compose:2.2.2")
+    implementation("com.github.skydoves:landscapist-coil:1.5.2")
+
+    //Lottie
+    implementation("com.airbnb.android:lottie-compose:5.2.0")
 
     //Dagger-Hilt
     implementation("com.google.dagger:hilt-android:2.50")
@@ -51,15 +79,20 @@ dependencies {
     implementation("androidx.hilt:hilt-navigation-compose:1.0.0")
     implementation("androidx.navigation:navigation-compose:2.5.3")
 
+    //ViewModel Lifecycle
+    implementation("androidx.lifecycle:lifecycle-viewmodel-ktx:$2.4.0")
+
+    //Navigation
+    implementation("androidx.navigation:navigation-compose:2.6.0")
+
+    //Youtube Player
+    implementation("com.pierfrancescosoffritti.androidyoutubeplayer:core:11.1.0")
+
     //Paging
     implementation ("androidx.paging:paging-runtime-ktx:3.2.1")
     implementation ("androidx.paging:paging-compose:3.2.1")
 
-    //Room
-    implementation("androidx.room:room-paging:2.6.1")
-    implementation ("androidx.room:room-ktx:2.6.1")
-    ksp("androidx.room:room-compiler:2.6.1")
-
+    implementation(project(":domain"))
     implementation(project(":util"))
 
     testImplementation("junit:junit:4.13.2")
