+++ ui/src/main/java/com/example/ui/component/CompErrorMessage.kt
@@ -1,0 +1,52 @@
+package com.example.ui.component
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.aspectRatio
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.airbnb.lottie.compose.LottieAnimation
+import com.airbnb.lottie.compose.LottieCompositionSpec
+import com.airbnb.lottie.compose.LottieConstants
+import com.airbnb.lottie.compose.rememberLottieComposition
+import com.example.ui.R
+
+@Composable
+fun CompErrorMessage(
+    modifier: Modifier = Modifier,
+    message: String
+) {
+    Column(
+        modifier = modifier
+            .fillMaxSize()
+            .verticalScroll(rememberScrollState()),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center
+    ) {
+        val composition by rememberLottieComposition(spec = LottieCompositionSpec.RawRes(R.raw.anim_error))
+        Row {
+            Spacer(modifier = Modifier.weight(3f))
+            LottieAnimation(
+                modifier = Modifier
+                    .weight(4f)
+                    .aspectRatio(1f),
+                composition = composition,
+                iterations = LottieConstants.IterateForever
+            )
+            Spacer(modifier = Modifier.weight(3f))
+        }
+
+        Spacer(modifier = Modifier.height(12.dp))
+        Text(text = message)
+    }
+}
