+++ ui/src/main/java/com/example/ui/feature/detailmovie/model/MovieVideoState.kt
@@ -1,0 +1,20 @@
+package com.example.ui.feature.detailmovie.model
+
+import com.example.domain.model.MovieVideoModel
+
+data class MovieVideoState(
+    val isLoading: Boolean = true,
+    val errorMessage: String = "",
+    val result: MovieVideoModel = MovieVideoModel(
+        site = "",
+        size = 0,
+        iso31661 = "",
+        name = "",
+        official = false,
+        id = "",
+        type = "",
+        publishedAt = "",
+        iso6391 = "",
+        key = ""
+    )
+)
